requirements.txt
python-dotenv==1.0.0
google-generativeai==0.8.3
html2text==2024.2.26
